+++ rhttp/lib/src/rust/frb_generated.io.dart
@@ -177,6 +177,9 @@ abstract class RustLibApiImplPlatform extends BaseApiImpl<RustLibWire> {
   ClientSettings dco_decode_box_autoadd_client_settings(dynamic raw);
 
   @protected
+  CookieSettings dco_decode_box_autoadd_cookie_settings(dynamic raw);
+
+  @protected
   HttpBody dco_decode_box_autoadd_http_body(dynamic raw);
 
   @protected
@@ -214,6 +217,9 @@ abstract class RustLibApiImplPlatform extends BaseApiImpl<RustLibWire> {
 
   @protected
   ClientSettings dco_decode_client_settings(dynamic raw);
+
+  @protected
+  CookieSettings dco_decode_cookie_settings(dynamic raw);
 
   @protected
   CustomProxy dco_decode_custom_proxy(dynamic raw);
@@ -312,6 +318,9 @@ abstract class RustLibApiImplPlatform extends BaseApiImpl<RustLibWire> {
 
   @protected
   ClientSettings? dco_decode_opt_box_autoadd_client_settings(dynamic raw);
+
+  @protected
+  CookieSettings? dco_decode_opt_box_autoadd_cookie_settings(dynamic raw);
 
   @protected
   HttpBody? dco_decode_opt_box_autoadd_http_body(dynamic raw);
@@ -508,6 +517,10 @@ abstract class RustLibApiImplPlatform extends BaseApiImpl<RustLibWire> {
       SseDeserializer deserializer);
 
   @protected
+  CookieSettings sse_decode_box_autoadd_cookie_settings(
+      SseDeserializer deserializer);
+
+  @protected
   HttpBody sse_decode_box_autoadd_http_body(SseDeserializer deserializer);
 
   @protected
@@ -551,6 +564,9 @@ abstract class RustLibApiImplPlatform extends BaseApiImpl<RustLibWire> {
 
   @protected
   ClientSettings sse_decode_client_settings(SseDeserializer deserializer);
+
+  @protected
+  CookieSettings sse_decode_cookie_settings(SseDeserializer deserializer);
 
   @protected
   CustomProxy sse_decode_custom_proxy(SseDeserializer deserializer);
@@ -653,6 +669,10 @@ abstract class RustLibApiImplPlatform extends BaseApiImpl<RustLibWire> {
 
   @protected
   ClientSettings? sse_decode_opt_box_autoadd_client_settings(
+      SseDeserializer deserializer);
+
+  @protected
+  CookieSettings? sse_decode_opt_box_autoadd_cookie_settings(
       SseDeserializer deserializer);
 
   @protected
@@ -881,6 +901,10 @@ abstract class RustLibApiImplPlatform extends BaseApiImpl<RustLibWire> {
       ClientSettings self, SseSerializer serializer);
 
   @protected
+  void sse_encode_box_autoadd_cookie_settings(
+      CookieSettings self, SseSerializer serializer);
+
+  @protected
   void sse_encode_box_autoadd_http_body(
       HttpBody self, SseSerializer serializer);
 
@@ -931,6 +955,10 @@ abstract class RustLibApiImplPlatform extends BaseApiImpl<RustLibWire> {
   @protected
   void sse_encode_client_settings(
       ClientSettings self, SseSerializer serializer);
+
+  @protected
+  void sse_encode_cookie_settings(
+      CookieSettings self, SseSerializer serializer);
 
   @protected
   void sse_encode_custom_proxy(CustomProxy self, SseSerializer serializer);
@@ -1041,6 +1069,10 @@ abstract class RustLibApiImplPlatform extends BaseApiImpl<RustLibWire> {
   @protected
   void sse_encode_opt_box_autoadd_client_settings(
       ClientSettings? self, SseSerializer serializer);
+
+  @protected
+  void sse_encode_opt_box_autoadd_cookie_settings(
+      CookieSettings? self, SseSerializer serializer);
 
   @protected
   void sse_encode_opt_box_autoadd_http_body(